Completely overhauled all the scenes
Added in final models with their materials and shaders into each scene, then cleaned up some objects that were no longer needed. Such as exit door, which is not part of melodydoor. Respective scripts with altered to make them work with the models new params and roles. Fixed a bug with limb light object to work with new limb systems.