Bug fixes to do with rotatable and melody door
a rotable fix to prevent multiple players from interacting with the same object led to a new visual bug occuring. This was resolved by adding a simple check. In addition, a melody door layer issue led to a bug with the player trying to rotate it instead of reseting the input code.