Added functionality to the rotate state and animation fixes
Added the ability to cancel the rotation of the object with the same key that launched that state, not just the pick up button, thereby making it more intuitive. Then fixed an animation issue where moving while in the process of going from movement to rotate state while holding an object would lead to a moving animation being played incorrectly.