Rewrite of exit light beam on light rediret object
A bug was created from a recent improvement on this object where, upon an exit of the connected light beam, a raycast is performed looking for any other beams that have connected on to the object previously. The bug was that the raycast wasnt protected from detecting the exitng beam itself, leading to a false connection during the rotation of that exiting light beam.